+++ src/Service/Translator/LegacyTranslator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Aeliot\Bundle\TransMaintain\Service;
+namespace Aeliot\Bundle\TransMaintain\Service\Translator;
 
 use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;